Implement forward function of sonnet_generation and paraphrase_detection
paraphrase_detection: logits are calculated using paraphrase_detection_head with the last token. Also labels are either 3919 (index of word "yes") or 8505 (index of word "no"), so we need to map them to 1 and 0 (map_labels function) before passing them to cross_entropy.

sonnet_generation: logits are for each token in the last hidden state.